+++ venus/src/layout/sidebar/components/sidebar-item/SidebarItemSubmenuLink.tsx
@@ -1,5 +1,5 @@
 import { NavLink } from "react-router-dom";
-import { SidebarSubmenuLink } from "../../model/link";
+import { SidebarSubmenuLink } from "../../../../model/interface/sidebar/link";
 
 interface SidebarItemSubmenuProps {
   link: SidebarSubmenuLink;
